update file paths
