+++ src/Components/Dropdowns/HeuristicDropdown.tsx
@@ -1,4 +1,4 @@
-import { Heuristic, changeHeuristic, heuristics } from "@/features/Visualizer";
+import { Heuristic, changeHeuristic, heuristics } from "@/features/Visualizer/visualizerSlice";
 import { useAppDispatch, useAppSelector } from "@/hooks";
 import { Dropdown, Navbar } from "@nextui-org/react";
