+++ src/Components/Dropdowns/PatternDropdown.tsx
@@ -1,4 +1,4 @@
-import { Pattern, changePattern, patterns } from "@/features/Visualizer";
+import { Pattern, changePattern, patterns } from "@/features/Visualizer/visualizerSlice";
 import { useAppDispatch, useAppSelector } from "@/hooks";
 import { Dropdown, Navbar } from "@nextui-org/react";
