+++ src/Components/Switchs/IsBidirectionalSwitch.tsx
@@ -1,4 +1,4 @@
-import { setIsBirectional } from "@/features/Visualizer";
+import { setIsBirectional } from "@/features/Visualizer/visualizerSlice";
 import { useAppDispatch, useAppSelector } from "@/hooks";
 import { Switch, SwitchEvent } from "@nextui-org/react";
 import { GiDirectionSign } from "react-icons/gi";
