+++ src/algorithms/Searching/Heuristics.ts
